Write the FunctionalController class
The trait, Controller, specifies the behavior for all concrete
Controller instances (the method handleRequest).

The class, FunctionalController, is a concrete instance of the
Controller. Additionally, it plays the role of AbstractClass in the
Template Method design pattern. In this role, it implements
handleRequest as a Template Method, implemented in terms of the
Primitive Operation, doRequest.

However, instead of writing a ConcreteClass to implement
PrimitiveOperation, this functional solution, asks clients to supply the
concrete implementation of PrimitiveOperation, doRequest. Consequently,
we have no need to write a ConcreteClass.